+++ 2D/Wave2.cpp
@@ -10,17 +10,17 @@ class Wave2 : public vb::Coloring {
 
   auto c(cpx z) -> vb::Color {
     double o = z.real() - M_PI / 2;
-    for (int i = 1; i < m; ++i) {
-      for (int j = 1; j < m; ++j) { o += coef[i][j] * sin(i * z.real()) * sin(j * z.imag()); }
+    for (size_t i = 1; i < m; ++i) {
+      for (size_t j = 1; j < m; ++j) { o += coef[i][j] * sin(i * z.real()) * sin(j * z.imag()); }
     }
     return vb::Indexed (o > 0 ? 2 : 1);
   }
 
   public:
   Wave2(const std::string &t, size_t n, size_t m_) : vb::Coloring(t, cpx(0, 0), cpx(M_PI, M_PI), n, [&](cpx z) { return c(z); }), m(m_) {
-    for (int i = 0; i < m; ++i) {
+    for (size_t i = 0; i < m; ++i) {
       coef.push_back({});
-      for (int j = 0; j < m; ++j) { coef.back().push_back(vb::prng.gaussian() / sqrt (i*i + j*j + 1)); }
+      for (size_t j = 0; j < m; ++j) { coef.back().push_back(vb::prng.gaussian() / sqrt(double(i)*i + double(j)*j + 1.0)); }
     }
   }
 };
